feat(routing): matchRoute matched result type uses supplied route types instead of AnyRoute
Previously the matchRoute function returned a union type with one member representing a successful
match, on that member there were event/route properties but they were typed using AnyRoute meaning
all type information was discarded. This was fine originally since it was only intended for internal
use, where that is the best level of type safety we can get anyway. But now that it is a public
function it should provide better type safety when the types of the routes to match are known. Now
it types the route/event properties correctly using the routing generics the function accepts