feat(xstate): update builders to use typegen states
Currently, the only way of specifying states is via type states, which are not used with typegen.

This means that the argument for `inState` ends up being `any` which is not great.

This commit updates the builder functions to extract the typegen'd states from the supplied state machine if it is using typegen, reverting to type states otherwise.

So now the `inState` arguments to the selectors/views will now be typegen aware.