fix(perf): don't delay routing events
This was like this due to an older performance issue where the popstate handler (that this code runs in) was taking a long time leading to UI lag. So the processing of the events and subsequent react re-render was moved to the next tick.

However the delay introduced by moving it to the next tick is actually worse