fix(routing): missing active route events
In certain situations (I don't know what, can reproduce with our production app integration tests, but not with a test in the project) the newly invoked slot machine will be rendered before the active routing event update has flowed through to the react context it's stored in, leading to them missing the routing event.

This would have been caused by the removal of the setTimeout wrapping the dispatching of routing events in fb898fccf52f0349ee702b2a18727bd6d616450c

So instead of storing it in component state, use a mutable ref and store that ref in the context so it can be updated immediately in handleRouteChange. It's only ever read when a view is mounted so don't need to cause re-renders of context consumers when it changes