feat(routing): composable dynamic routes
Overhauls the route creation system to enable composing dynamic routes together. With the previous system only static routes (a fixed url) could be composed together. If you wanted a dynamic URL you were limited to not creating sub routes from it.

Now everything is an equivalent to a dynamic route and the static routes are built ontop of them.

The `any` type massacre in the create route functions has only gotten worse, sadly.