fix(routing): notify about double routing roots
Having two routing roots on the page at once causes problems, as they are both listening to the same routing events

If they have the same routes supplied it's probably ok, just double broadcast messages.

But if they have different routes then one will broadcast a 404 event, and one will not.

Better not to do it at all. So that will now throw an error in dev and log an error message in prod to enable it to be caught