feat(routing): forwards refs for Link component
Moving preventDefault down to just before the navigation event

In the onClick event Radix is checking whether the default event was prevented and bails out of doing its things.

Since this should just be for preventing the browser from following the link's url, it should be fine to run it just before the navigate call.